+++ main/cvar.c
@@ -106,7 +106,7 @@ cvar_t *cvar_find(char *cvar_name)
 char *cvar_complete(char *text)
 {
 	cvar_t *ptr;
-	int len = strlen(text);
+	int len = (int)strlen(text);
 
 	if (!len)
 		return NULL;
